requirements.txt
pytorch_lightning==2.2.5
numpy==1.24.4
pandas==1.5.3
h5py>=3.11.0
PYyaml==6.0.1
scikit-image==0.23.2
jsonargparse[signatures]>=4.27.7
tensorboard==2.17.0
tensorboardX==2.6.2.2